% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggframe.R
\name{facet}
\alias{facet}
\title{facet}
\usage{
facet(.data, facet, ...)
}
\arguments{
\item{.data}{data.frame or ggframe}

\item{facet}{a string so that \verb{facet_<facet>} will be called}

\item{...}{passed to \verb{facet_<facet>()}}
}
\description{
facet
}
