% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggframe.R
\name{geom}
\alias{geom}
\title{Add a geom}
\usage{
geom(data, geom, ...)
}
\arguments{
\item{data}{data.frame or ggframe}

\item{geom}{a string so that \verb{geom_<geom>} will be called}

\item{...}{passed to \verb{geom_<geom>()}}
}
\description{
Add a geom
}
