% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggframe.R
\name{print.ggframe}
\alias{print.ggframe}
\title{print a ggframe}
\usage{
\method{print}{ggframe}(x, plot = TRUE, ...)
}
\arguments{
\item{x}{x}

\item{plot}{whether to plot or print data.frame and attributes}

\item{...}{for compatibility with other methods}
}
\description{
print a ggframe
}
