% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggframe.R
\name{set_coord}
\alias{set_coord}
\title{set coord}
\usage{
set_coord(data, coord, ...)
}
\arguments{
\item{data}{data.frame or ggframe}

\item{...}{passed to \verb{geom_<geom>()}}

\item{geom}{a string so that \verb{geom_<geom>} will be called}
}
\description{
set coord
}
