% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggframe.R
\name{set_guide}
\alias{set_guide}
\title{set guide}
\usage{
set_guide(.data, guide, ...)
}
\arguments{
\item{.data}{data.frame or ggframe}

\item{guide}{a string so that \verb{guide_<guide>} will be called}

\item{...}{passed to \verb{guide_<guide>()}}
}
\description{
set guide
}
