% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggframe.R
\name{set_labs}
\alias{set_labs}
\title{set labs}
\usage{
set_labs(
  .data,
  ...,
  title = ggplot2::waiver(),
  subtitle = ggplot2::waiver(),
  caption = ggplot2::waiver(),
  tag = ggplot2::waiver()
)
}
\arguments{
\item{.data}{data.frame or ggframe}

\item{...}{A list of new name-value pairs. The name should be an aesthetic.}

\item{title}{The text for the title.}

\item{subtitle}{The text for the subtitle for the plot which will be
displayed below the title.}

\item{caption}{The text for the caption which will be displayed in the
bottom-right of the plot by default.}

\item{tag}{The text for the tag label which will be displayed at the
top-left of the plot by default.}
}
\description{
set labs
}
