% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggframe.R
\name{set_coord}
\alias{set_coord}
\title{set coord}
\usage{
set_coord(.data, coord, ...)
}
\arguments{
\item{.data}{data.frame or ggframe}

\item{coord}{a string so that \verb{coord_<coord>} will be called}

\item{...}{passed to \verb{coord_<coord>()}}
}
\description{
set coord
}
