% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggframe.R
\name{set_scale}
\alias{set_scale}
\title{set scale}
\usage{
set_scale(.data, scale, ...)
}
\arguments{
\item{.data}{data.frame or ggframe}

\item{scale}{a string so that \verb{scale_<scale>} will be called}

\item{...}{passed to \verb{scale_<scale>()}}
}
\description{
set scale
}
