% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggframe.R
\name{set_theme}
\alias{set_theme}
\title{set theme}
\usage{
set_theme(.data, theme, ...)
}
\arguments{
\item{.data}{data.frame or ggframe}

\item{theme}{a string so that \verb{theme_<theme>} will be called}

\item{...}{passed to \verb{theme_<theme>()}}
}
\description{
set theme
}
